+++ arduino/Esp32Mousr/include/controller.h
@@ -33,6 +33,11 @@ private:
     Controller();
 
 public:
+    /**
+     * @brief Get the singleton controller instance
+     * 
+     * @return Controller* 
+     */
     static Controller *getInstance()
     {
         if (singleton == NULL)
@@ -43,12 +48,31 @@ public:
         return singleton;
     };
 
+    /**
+     * @brief Starts the controller
+     * 
+     * @param irqPin Interrupt pin
+     * @param addr i2c address
+     * @return true Controller successfully started
+     * @return false Controller could not start
+     */
     bool begin(uint8_t irqPin,
                uint8_t addr = 0x49);
 
     bool end();
 
+    /**
+     * @brief Get the analog stick peripheral
+     * 
+     * @return ControllerAnalogStick* 
+     */
     ControllerAnalogStick *getStick();
+
+    /**
+     * @brief Get the controller buttons peripheral
+     * 
+     * @return ControllerButtons* 
+     */
     ControllerButtons *getButtons();
 
 private:
