+++ arduino/Esp32Mousr/platformio.ini
@@ -28,6 +28,7 @@ src_filter =
 	+<controller.cpp>
 	+<main.cpp>
 	+<oled.cpp>
+	+<oleddraw.cpp>
 	+<utility.cpp>
 	+<buttons.cpp>
 	+<analog.cpp>
